requirement.txt
pytorch
json
tqdm
matplotlib
torchmetrics
torchaudio
os
